+++ packages/client/src/reexports/polkadot-sdk-compat.ts
@@ -1,2 +1,1 @@
 export * from "@polkadot-api/polkadot-sdk-compat"
-export { default } from "@polkadot-api/polkadot-sdk-compat"